fix(core): extract getInstalledNxVersion to shared util
Move `getInstalledNxVersion` (and its cache-shielded resolver helper) out
of the daemon directory into `utils/installed-nx-version.ts` so it can
be used by non-daemon code. `is-nx-version-mismatch.ts` now re-exports
it for backward compatibility on the daemon import path.

Replace the bespoke `getNxVersion` in `set-up-ai-agents.ts` with the
shared `getInstalledNxVersion`, falling back to the workspace
package.json declaration and a default only when nothing is actually
installed. Removes a ~35-line duplicate of resolver logic that was
itself the original source of the cache-pollution chain in #35444.
